+++ t/utils.lisp
@@ -1,0 +1,36 @@
+(in-package :cl-user)
+(defpackage hamcrest.t.utils
+  (:use :cl
+        :prove
+        :hamcrest.utils))
+(in-package :hamcrest.t.utils)
+
+
+(plan 1)
+
+(subtest
+    "Checking alistp predicate"
+  (let ((good-values (list nil
+                           '((:foo . 1)
+                             (:bar . 2))))
+        (bad-values (list 100500
+                          "Blah"
+                          '((:foo . 1)
+                            :bar 2)
+                          #(#(:foo 1)
+                            #(:bar 2)))))
+    (loop
+       :for value
+       :in good-values
+       :do (is (alistp value)
+               t
+               (format nil "~S is a proper alist" value)))
+
+    (loop
+       :for value
+       :in bad-values
+       :do (is (alistp value)
+               nil
+               (format nil "~S is not a proper alist" value)))))
+
+(finalize)